error: 결과값 6이 아닌 0출력
